plot.new() has no argument anymore
git-svn-id: https://svn.r-project.org/R/trunk@11112 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/include/R_ext/Mathlib.h
@@ -554,7 +554,7 @@ double	chebyshev_eval(double, double *, int);
 
 void	gammalims(double*, double*);
 double	lgammacor(double);/* log(gamma) correction */
-double  stirlerr(int);/* Stirling expansion "error"  */
+double  stirlerr(double);/* Stirling expansion "error"  */
 double	gammafn(double);
 double	gamma_cody(double);
 double	lgammafn(double);
